p1/image/requirements.txt
numpy>=1.23.5,<2
torch==2.2.2
torchvision==0.17.2
Pillow>=10.2.0
